+++ CHANGELOG.md
@@ -19,6 +19,7 @@ to [sourceforge feature requests](https://sourceforge.net/p/jabref/features/) by
 - Changes in customized entry types are now directly reflected in the table when clicking "Apply" or "OK"
 - Reference list generation works for OpenOffice/LibreOffice again, fixes #593
 - ACM fetcher works again, fixes #545
+- Use default file icon for custom external file types, fixes #598
 
 ### Removed